ruff format

+++ database_func/action_on_admin.py
@@ -8,25 +8,15 @@ class ActionsOnAdmin:
     async def add_admin(future_admin: dict[str, Any]) -> None:
         await admins_db.connect_async(reuse_if_open=True)
         admins_db.create_tables([AdminUsers])
-        (
-            AdminUsers
-            .insert(future_admin)
-            .execute()
-        )
+        (AdminUsers.insert(future_admin).execute())
         admins_db.commit()
         await admins_db.close_async()
-
 
     @staticmethod
     async def del_admin(ex_admin: dict[str, Any]) -> None:
         await admins_db.connect_async(reuse_if_open=True)
         admins_db.create_tables([AdminUsers])
-        (
-            AdminUsers
-            .delete()
-            .where(AdminUsers.id == ex_admin["id"])
-            .execute()
-        )
+        (AdminUsers.delete().where(AdminUsers.id == ex_admin["id"]).execute())
         admins_db.commit()
         await admins_db.close_async()
 
@@ -40,8 +30,10 @@ class ActionsOnAdmin:
                 admins_id: list[int] = [admin.id for admin in data]
                 return admins_id
             case False:
-                admins_data: list[tuple] = [(admin.id, admin.first_name, admin.last_name, admin.username)
-                                     for admin in data]
+                admins_data: list[tuple] = [
+                    (admin.id, admin.first_name, admin.last_name, admin.username)
+                    for admin in data
+                ]
                 return admins_data
         admins_db.commit()
         await admins_db.close_async()
